requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4,<4.0
mysqlclient
django-redis
djangorestframework-simplejwt
django-debug-toolbar
celery[redis]==5.2.3
